+++ joss/paper.bib
@@ -6,7 +6,8 @@
   number={2},
   pages={341--363},
   year={1986},
-  publisher={SIAM}
+  publisher={SIAM},
+  doi={10.1109/sfcs.1983.11},
 }
 
 @inproceedings{murali1997consistent,
@@ -14,7 +15,8 @@
   author={Murali, TM and Funkhouser, Thomas A},
   booktitle={Proceedings of the Symposium on Interactive 3D Graphics},
   pages={155--162},
-  year={1997}
+  year={1997},
+  doi={10.1145/253284.253326},
 }
 
 @techreport{hagberg2008exploring,
@@ -24,7 +26,8 @@
   place        = {United States},
   organization = {Los Alamos National Laboratory (LANL), Los Alamos, NM (United States)},
   year         = {2008},
-  month        = {01}
+  month        = {01},
+  doi          = {10.25080/tcwv9851},
 }
 
 @article{nan2021easy3d,
@@ -36,7 +39,7 @@
   number = {64},
   pages = {3255},
   doi = {10.21105/joss.03255},
-  url = {https://doi.org/10.21105/joss.03255}
+  url = {https://doi.org/10.21105/joss.03255},
 }
 
 @article{chen2022points2poly,
@@ -46,9 +49,9 @@
   pages = {58-73},
   year = {2022},
   issn = {0924-2716},
-  doi = {https://doi.org/10.1016/j.isprsjprs.2022.09.017},
+  doi = {10.1016/j.isprsjprs.2022.09.017},
   url = {https://www.sciencedirect.com/science/article/pii/S0924271622002611},
-  author = {Zhaiyu Chen and Hugo Ledoux and Seyran Khademi and Liangliang Nan}
+  author = {Zhaiyu Chen and Hugo Ledoux and Seyran Khademi and Liangliang Nan},
 }
 
 @article{chen2024polygnn,
@@ -58,7 +61,7 @@
   pages = {693-706},
   year = {2024},
   issn = {0924-2716},
-  doi = {https://doi.org/10.1016/j.isprsjprs.2024.09.031},
+  doi = {10.1016/j.isprsjprs.2024.09.031},
   url = {https://www.sciencedirect.com/science/article/pii/S0924271624003691},
   author = {Zhaiyu Chen and Yilei Shi and Liangliang Nan and Zhitong Xiong and Xiao Xiang Zhu},
 }
@@ -71,7 +74,8 @@
   publisher="Springer Nature Switzerland",
   address="Cham",
   pages="357--373",
-  isbn="978-3-031-72904-1"
+  isbn="978-3-031-72904-1",
+  doi="10.1007/978-3-031-72904-1_21",
 }
 
 @manual{sage2024sage,
