+++ tests/test_complex.py
@@ -1,102 +1,219 @@
-from pathlib import Path
-
+import pytest
 import numpy as np
-import trimesh
+from pathlib import Path
+from abspy.complex import CellComplex
 
-from abspy import CellComplex
-
-dir_tests = Path(__file__).parent
-
-
-def example_cell_complex_from_planes():
-    """
-    Simple CellComplex construction from specified planes and bounds.
-    """
+@pytest.fixture
+def simple_planes():
+    """Create planes for testing using a verified working configuration."""
     # start from two planes
     planes = np.array([[0, 1, 0, -50], [0, 0, 1, -50]])
 
     # specify the bounds
-    bounds = np.array([[[0, 50, 0], [100, 50, 100]], [[0, 0, 50], [100, 50, 50]]])
+    aabbs = np.array([[[0, 50, 0], [100, 50, 100]], [[0, 0, 50], [100, 50, 50]]])
+    
+    # Create arbitrary OBBs for testing
+    obbs = np.array([
+        [[0, 50, 0], [100, 50, 0], [100, 50, 100], [0, 50, 100]],   # First plane
+        [[0, 0, 50], [100, 0, 50], [100, 50, 50], [0, 50, 50]]      # Second plane
+    ])
+    
+    return planes, aabbs, obbs
 
+
+@pytest.fixture
+def simple_complex(simple_planes):
+    """Create a simple cell complex."""
+    planes, aabbs, obbs = simple_planes
+    # Create points for each plane (simple grid of points)
+    points = np.array([
+        np.array([[x, 50, z] for x in np.linspace(0, 100, 5) for z in np.linspace(0, 100, 5)]),
+        np.array([[x, y, 50] for x in np.linspace(0, 100, 5) for y in np.linspace(0, 50, 5)])
+    ], dtype=object)
+    
     # specify the initial bound
-    initial_bound = [[0, 0, 0], [100, 100, 100]]
-
-    # initialise the cell complex
-    cell_complex = CellComplex(planes, bounds, initial_bound=initial_bound, build_graph=True)
-
-    # prioritise certain (vertical by default) planes
-    cell_complex.prioritise_planes()
-
-    # construct the complex
-    cell_complex.construct()
-
-    # print out info
-    cell_complex.print_info()
-
-    # boundary cells
-    cells_boundary = cell_complex.cells_boundary()
-    print(f'boundary cells: {cells_boundary}')
-
-    # cell representatives
-    num_representatives = 1000
-    representatives = cell_complex.cell_representatives(location='skeleton', num=num_representatives)
-    representatives = np.concatenate(representatives, axis=0)
-    colours = np.zeros([num_representatives * cell_complex.num_cells, 4])
-    for i in range(cell_complex.num_cells):
-        colours[num_representatives * i: num_representatives * (i + 1)] = trimesh.visual.color.random_color()
-
-    # visualise cell representatives
-    representatives_vis = trimesh.PointCloud(representatives, colours)
-    representatives_vis.export(dir_tests / 'test_output' / 'test_vis.ply', file_type='ply')
-    try:
-        representatives_vis.show()
-    except (Exception, ImportError) as e:
-        print(f'visualization skipped: {e}')
-
-    # save cell complex CC file
-    cell_complex.save(dir_tests / 'test_output' / 'complex.cc')
-
-    # save cells to OBJ and PLM files
-    cell_complex.save_obj(dir_tests / 'test_output' / 'cells.obj', use_mtl=True)
-    cell_complex.save_plm(dir_tests / 'test_output' / 'cells.plm')
+    initial_bound = np.array([[0, 0, 0], [100, 100, 100]])
+        
+    return CellComplex(planes, aabbs, obbs=obbs, points=points, 
[…]
