### 2025-03-25 #7946 @editorialbot
[REVIEW]: abspy: A Python package for 3D adaptive binary space partitioning and modeling
**Submitting author:** @chenzhaiyu (<a href="http://orcid.org/0000-0001-7084-0994">Zhaiyu Chen</a>)
**Repository:** https://github.com/chenzhaiyu/abspy
**Branch with paper.md** (empty if default branch): joss-stable
**Version:** v0.3.0
