### 2025-03-26 chenzhaiyu/abspy#24 @preetishkakkar
Error when trying to install conda packages
Part of review for https://github.com/openjournals/joss-reviews/issues/7946

**Describe the bug**
conda env create -f environment.yml gives an error saying

Solving environment: failed
Channels:
 - conda-forge
 - defaults
Platform: win-64
Solving environment: failed

LibMambaUnsatisfiableError: Encountered problems while solving:
  - nothing provides cypari2 needed by sage-10.2-hd8ed1ab_0

Could not solve for environment specs
The following package could not be installed
└─ sage 10.2**  is not installable because it requires
   └─ cypari2, which does not exist (perhaps a missing channel).

**To Reproduce**
Steps to reproduce the behavior:
1.) Download and install AnaConda (latest version is 24.9.2)
2.) Run conda env create -f environment.yml
