### 2025-03-26 chenzhaiyu/abspy#24 @chenzhaiyu
Hi @preetishkakkar, thank you for reporting the error. I've noticed that your environment is on Windows. As stated in our [FAQ](https://github.com/chenzhaiyu/abspy#faq), abspy does not support Windows natively because it depends on [SageMath](https://www.sagemath.org/index.html), which goes native only with Linux and macOS.
> For Windows users, you may need to build [SageMath from source](https://doc.sagemath.org/html/en/installation/source.html) or install all other dependencies into a [pre-built SageMath environment](https://doc.sagemath.org/html/en/installation/binary.html). Otherwise, virtualization with [docker](https://www.docker.com/) may come to the rescue.

Would you be able to try one of the above suggested moves?
