### 2025-03-28 chenzhaiyu/abspy#25 @preetishkakkar
README.md file contains Quick Start examples which refer to various files such as points.vg and mesh.obj (that don't exists)
Part of the review for https://github.com/openjournals/joss-reviews/issues/7946

**Describe the bug**
The readme file contains Quick Start examples which refer to various files such as points.vg and mesh.obj, however, it's not clear where to get these files from. The repo has a tutorial folder containing various tutorials along with test_point.vg etc file in the data folder. I think it would be useful to have a readme file inside the tutorial folder referenced by a top-level readme file.
