### 2025-03-30 chenzhaiyu/abspy#26 @preetishkakkar
parallelize some parts of code
Part of the review for https://github.com/openjournals/joss-reviews/issues/7946

Currently, a lot of code inside primitive.py (such as get_points/primitives) is sequential, I would like to suggest parallelizing some of this code to make these operations more performant.
