### 2025-04-06 chenzhaiyu/abspy#27 @preetishkakkar
Remove NumPy print options from logging configuration (nit pick)
Part of the review for openjournals/joss-reviews#7946

The current implementation of attach_to_log() function changes global NumPy print settings by calling np.set_printoptions(precision=5, suppress=True).

 This creates several issues:

* It's usually not a good idea for logging configuration function to modify NumPy's global state
* It can cause unexpected side effects, the Users of the logging function may not expect or want their NumPy print options to be changed
* Different parts of the application may require different NumPy print formatting
* Since we are changing global state, it will affect the entire application, not just the logged output.

Is there a better solution to this? this is just nitpick
