### 2025-04-06 chenzhaiyu/abspy#28 @preetishkakkar
Add more details on how this library compares to other commonly used libraries (in Paper)
Part of the review for https://github.com/openjournals/joss-reviews/issues/7946

The paper does reference prior work such as fixed-grid methods like voxels, and exhaustive space decomposition, however, it does not explicitly name or compare to other widely-used software packages that might offer similar 3D modeling, partitioning, or BSP-based functionality. It would be good to compare the package against some other libraries such as CGAL, Open3D or any others.
