### 2025-04-07 6e843ff
Remove numpy print options (#27)
