### 2025-04-07 chenzhaiyu/abspy#27 @chenzhaiyu
You're right, I just removed the unnecessary line :)
