### 2025-04-07 342fca1
Paper compare with CGAL and Open3D
