+++ joss/paper.bib
@@ -1,6 +1,6 @@
 @article{edelsbrunner1986constructing,
   title={Constructing arrangements of lines and hyperplanes with applications},
-  author={Edelsbrunner, Herbert and O’Rourke, Joseph and Seidel, Raimund},
+  author={Edelsbrunner, Herbert and O'Rourke, Joseph and Seidel, Raimund},
   journal={SIAM Journal on Computing},
   volume={15},
   number={2},
@@ -67,15 +67,15 @@
 }
 
 @inproceedings{sulzer2024concise,
-  author="Sulzer, Raphael and Lafarge, Florent",
-  title="Concise Plane Arrangements for Low-Poly Surface and Volume Modelling",
-  booktitle="Computer Vision -- ECCV 2024",
-  year="2024",
-  publisher="Springer Nature Switzerland",
-  address="Cham",
-  pages="357--373",
-  isbn="978-3-031-72904-1",
-  doi="10.1007/978-3-031-72904-1_21",
+  author={Sulzer, Raphael and Lafarge, Florent},
+  title={Concise Plane Arrangements for Low-Poly Surface and Volume Modelling},
+  booktitle={Computer Vision -- ECCV 2024},
+  year={2024},
+  publisher={Springer Nature Switzerland},
+  address={Cham},
+  pages={357--373},
+  isbn={978-3-031-72904-1},
+  doi={10.1007/978-3-031-72904-1_21},
 }
 
 @manual{sage2024sage,
@@ -85,4 +85,26 @@
   Title        = {{S}age {M}athematics {S}oftware ({V}ersion 10.2)},
   note         = {{\tt http://www.sagemath.org}},
   Year         = {2024},
+}
+
+@inproceedings{fabri2009cgal,
+  author = {Fabri, Andreas and Pion, Sylvain},
+  title = {{CGAL}: the Computational Geometry Algorithms Library},
+  year = {2009},
+  isbn = {9781605586496},
+  publisher = {Association for Computing Machinery},
+  address = {New York, NY, USA},
+  doi = {10.1145/1653771.1653865},
+  booktitle = {Proceedings of the 17th ACM SIGSPATIAL International Conference on Advances in Geographic Information Systems},
+  pages = {538–539},
+  numpages = {2},
+  location = {Seattle, Washington},
+  series = {GIS '09},
+}
+
+@article{zhou2018open3d,
+  title={{Open3D}: a modern library for {3D} data processing},
+  author={Zhou, Qian-Yi and Park, Jaesik and Koltun, Vladlen},
+  journal={arXiv preprint arXiv:1801.09847},
+  year={2018},
 }
