+++ joss/paper.md
@@ -23,7 +23,7 @@ bibliography: paper.bib
 
 # Statement of need
 
-Many applications—from remote sensing and robotics to computer graphics—require efficient methods for representing and analyzing complex 3D spaces. Fixed-grid (e.g., voxels) and exhaustive [@edelsbrunner1986constructing] approaches often struggle to capture the intricate geometry of data or incur prohibitive computational costs. Adaptive binary space partitioning [@murali1997consistent] overcomes these challenges by dynamically subdividing 3D space according to its intrinsic geometric features and constructing a plane arrangement that faithfully reflects the data structure, as shown in \autoref{fig:2d}. With its modern, user-friendly interface built on this adaptive mechanism, *abspy* enables accurate and efficient 3D modeling and analysis, allowing researchers to tackle large-scale and complex reconstruction challenges with great reliability and precision. Moreover, while other packages may offer sparse partitioning capabilities, *abspy* distinguishes itself through seamless integration with existing research workflows and enhanced interoperability, making it fully suitable for comprehensive 3D modeling.
+Many applications—from remote sensing and robotics to computer graphics—require efficient methods for representing and analyzing complex 3D spaces. Fixed-grid (e.g., voxels) and exhaustive [@edelsbrunner1986constructing] approaches often struggle to capture the intricate geometry of data or incur prohibitive computational costs. Adaptive binary space partitioning [@murali1997consistent] overcomes these challenges by dynamically subdividing 3D space according to its intrinsic geometric features and constructing a plane arrangement that faithfully reflects the data structure, as shown in \autoref{fig:2d}. With its modern, user-friendly interface built on this adaptive mechanism, *abspy* enables accurate and efficient 3D modeling and analysis, allowing researchers to tackle large-scale and complex reconstruction challenges with great reliability and precision. Moreover, unlike geometric libraries such as CGAL [@fabri2009cgal] and Open3D [@zhou2018open3d], which primarily offer versatile low-level tools, abspy uniquely integrates adaptive binary space partitioning and exact arithmetic within a high-level Python-native framework, facilitating direct integration into research workflows tailored specifically for compact and dynamic 3D modeling tasks.
 
 ![A 2D illustration for adaptive binary space partitioning.\label{fig:2d}](assets/2d.png){ width=85% }
 
