### 2025-04-07 chenzhaiyu/abspy#28 @chenzhaiyu
Hi @preetishkakkar, the main distinction is that libraries like CGAL primarily offer versatile low-level tools (similar to SageMath in our case); they would still require significant effort from users to assemble the higher-level functionalities that abspy readily provides. Additionally, I believe Open3D lacks an exact arithmetic kernel to handle degenerate geometric cases. I would expect readers recognize abspy as a specialized tool tailored specifically for this task. I've concisely incorporated your suggestion into the paper: https://github.com/chenzhaiyu/abspy/commit/342fca12fce89e3fb55e87e943213e5e47c36fdb :)
