### 2025-04-20 dd9612c
Fix `environment.yml` (#29)
