### 2025-04-20 d1c9148
Enable matrix builds also for macOS in workflow
