### 2025-04-20 8c5a7cb
Update Conda setup to specify latest miniconda version
