### 2025-04-20 fbc8480
Set MPLBACKEND to unblock macOS builds
