### 2025-04-20 cf4c662
Fix obsolete benchmark (#30)
