### 2025-04-21 chenzhaiyu/abspy#31 @tuelwer
Paper improvements
This issue is part of my [review for JOSS](https://github.com/openjournals/joss-reviews/issues/7946). The paper is relatively short and can be improved at some points:

- [ ] The inclusion of additional background context could improve the clarity of the paper. It would be nice if a reader who is not familiar with algorithmic geometry/computer graphics could easily understand which problem this package solves. I acknowledge that Figure 1 and 2 already offer some insight but the paper could benefit from additional explanations and perhaps a clear problem definition. Enhancing the caption of Figure 2 could also help: what types of representations are being shown there?
- [x] I found an implementation of binary space partitioning in the [python-tcod library](https://python-tcod.readthedocs.io/en/latest/tcod/bsp.html). Please, consider citing this implementation if it is relevant to abspy. If relevant it would also be interesting to compare this implementation with the implementation in abspy.
- [x] The paper mentions that abspy has been used by Sulzer & Lafarge (2024). I was not able to verify that claim. Can you kindly point me to a citation or relevant lines of code?
- [x] Minor: Fix reference in line 16: put the citation after the word "approaches".
- [ ] Minor: Improve figure placement. There is a single line of text under Figure 1.

The current paper is 2 pages long. I know that JOSS papers can have at least up to 4 pages thus there is room to accommodate these requested changes.
