### 2025-04-21 chenzhaiyu/abspy#32 @tuelwer
Test coverage is somewhat low
Hi @chenzhaiyu, this is part of my https://github.com/openjournals/joss-reviews/issues/7946. I calculated the [coverage](https://en.wikipedia.org/wiki/Code_coverage) by the tests using the `coverage` package:
```
coverage run -m pytest
coverage report
```
Here are the results:
```
Name                      Stmts   Miss  Cover
---------------------------------------------
abspy/__init__.py             6      0   100%
abspy/complex.py            495    219    56%
abspy/graph.py              208    150    28%
abspy/logger.py              22      8    64%
abspy/primitive.py          504    289    43%
abspy/version.py              3      1    67%
tests/test_complex.py       105      0   100%
tests/test_graph.py          68      1    99%
tests/test_primitive.py     143     12    92%
---------------------------------------------
TOTAL                      1554    680    56%
```
In my opinion 56% is somewhat low. Could you add some tests to improve the coverage of your tests? I would suggest aiming for 80%.
