### 2025-04-21 chenzhaiyu/abspy#33 @tuelwer
Add some explanations for the tutorials
This issue is part of my https://github.com/openjournals/joss-reviews/issues/7946. Similar to #31 the tutorials would benefit from adding some additional context. What is happening in each tutorial on a high level? What is the expected output (maybe add a screenshot)? Is it possible to show intermediate steps? How do I inspect the output files? Maybe an additional page in the [docs](https://abspy.readthedocs.io/en/latest/) might be a good place to accommodate these changes.
