### 2025-04-23 9e0b221
More explanations for tutorials on the doc page (#33)
