+++ tutorials/README.md
@@ -1,18 +1,0 @@
-# Tutorials
-
-This directory contains self-explanatory tutorials to show the functionality of `abspy`.
-
-## Content
-
-- `tutorial_primitive.py`: Extracting primitives from point clouds (`.vg` and `.bvg` files) and reference meshes
-- `tutorial_complex.py`: Creating cell complexes from planes and meshes
-- `tutorial_graph.py`: Simple example of working with adjacency graphs
-- `tutorial_combined.py`: Combining multiple operations into a workflow
-
-- `data/`: Test data files used by the tutorials
-    - `.vg/.bvg`: Vertex group files containing point cloud data and planar primitives
-    - `.obj/.ply`: 3D model files
-    - `.adjlist`: Adjacency graph data
-    - `.npy`: NumPy array data
-
-For more detailed API information, please refer to the [abspy documentation](https://abspy.readthedocs.io/en/latest/api.html).
