### 2025-04-24 6c2b108
Improve paper background and position (#31)
