+++ joss/paper.md
@@ -19,15 +19,15 @@ bibliography: paper.bib
 
 # Summary
 
-*abspy* is a Python library for 3D adaptive binary space partitioning and modeling. At its core, *abspy* constructs a plane arrangement by recursively subdividing 3D space with planar primitives to form a linear cell complex that reflects the underlying geometric structure of the data. This adaptive approach iteratively refines the spatial decomposition, enabling efficient modeling and analysis. Built on robust computational techniques and data structures designed for interoperability, *abspy* offers a flexible framework that supports advanced operations such as surface reconstruction, volumetric analysis, and feature extraction for machine learning.
+Efficient and robust space partitioning of 3D space underpins many applications in computer graphics. *abspy* is a Python package for adaptive binary space partitioning and 3D modeling. At its core, *abspy* constructs a plane arrangement by recursively subdividing 3D space with planar primitives to form a linear cell complex that reflects the underlying geometric structure. This adaptive scheme iteratively refines the spatial decomposition, producing a compact representation that supports efficient 3D modeling and analysis. Built on robust computational techniques and interoperable data structures, *abspy* provides a flexible framework for advanced operations such as surface reconstruction, volumetric analysis, and feature extraction for machine learning.
 
 # Statement of need
 
-Many applications—from remote sensing and robotics to computer graphics—require efficient methods for representing and analyzing complex 3D spaces. Fixed-grid (e.g., voxels) and exhaustive [@edelsbrunner1986constructing] approaches often struggle to capture the intricate geometry of data or incur prohibitive computational costs. Adaptive binary space partitioning [@murali1997consistent] overcomes these challenges by dynamically subdividing 3D space according to its intrinsic geometric features and constructing a plane arrangement that faithfully reflects the data structure, as shown in \autoref{fig:2d}. With its modern, user-friendly interface built on this adaptive mechanism, *abspy* enables accurate and efficient 3D modeling and analysis, allowing researchers to tackle large-scale and complex reconstruction challenges with great reliability and precision. Moreover, unlike geometric libraries such as CGAL [@fabri2009cgal] and Open3D [@zhou2018open3d], which primarily offer versatile low-level tools, abspy uniquely integrates adaptive binary space partitioning and exact arithmetic within a high-level Python-native framework, facilitating direct integration into research workflows tailored specifically for compact and dynamic 3D modeling tasks.
[…]
