### 2025-04-24 chenzhaiyu/abspy#31 @chenzhaiyu
Thank you for the suggestions!

* I’ve added more background and clarified the problem definition in both the paper and #33 in the docs, and expanded each figure caption with more detailed descriptions.
* I checked _python-tcod_, and, since it only generates 2D dungeon layouts via axis-aligned BSP, I believe it isn’t directly relevant to our 3D, arbitrarily-oriented partitioning focus. Please let me know if you see any overlap I’ve missed.
* In Sec 4.1 of their [nice paper](https://arxiv.org/pdf/2404.06154) _abspy_ appears as 'P2P', and git history [here](https://github.com/raphaelsulzer/compod/commits/main/?author=chenzhaiyu).
* Reference position fixed.
* Figure placement improved too. :)
