### 2025-04-24 6b756f7
Improve test coverage (#32)
