+++ tests/test_complex.py
@@ -71,6 +71,147 @@ def test_construct(simple_complex):
     assert cc.num_planes == 2
 
 
+def test_refine_planes(simple_complex):
+    """Test plane refinement functionality."""
+    cc = simple_complex
+    
+    # Save original plane data
+    original_planes = cc.planes.copy()
+    original_aabbs = cc.aabbs.copy()
+    
+    # Refine planes with default parameters
+    cc.refine_planes()
+    
+    # Since our test data doesn't contain similar planes,
+    # the number of planes should remain the same
+    assert cc.planes.shape == original_planes.shape
+    assert cc.aabbs.shape == original_aabbs.shape
+
+
+def test_prioritise_planes(simple_complex):
+    """Test plane prioritization functionality."""
+    cc = simple_complex
+    
+    # Store original planes
+    original_planes = cc.planes.copy()
+    
+    # Prioritize planes (without verticals first)
+    cc.prioritise_planes(prioritise_verticals=False)
+    
+    # The total number should be the same
+    assert cc.planes.shape == original_planes.shape
+    assert (cc.planes == original_planes).all()  # no reorder
+    
+    # Test with vertical prioritization
+    cc = simple_complex
+    cc.prioritise_planes(prioritise_verticals=True)
+    assert cc.planes.shape == original_planes.shape
+    assert (cc.planes == original_planes).all()  # no reorder
+
+
+def test_sort_planes():
+    """Test plane sorting functionality."""
+    # Create planes with different sized bounds
+    planes = np.array([
+        [0, 0, 1, 0],
+        [0, 1, 0, 0],
+        [1, 0, 0, 0],
+    ])
+    
+    # Use different sized AABBs
+    aabbs = np.array([
+        [[-1, -1, -0.01], [1, 1, 0.01]],     # Small z-extent
+        [[-1, -0.01, -5], [1, 0.01, 5]],     # Large z-extent
+        [[-0.01, -3, -3], [0.01, 3, 3]],     # Medium extent
+    ])
+    
+    cc = CellComplex(planes, aabbs, quiet=True)
+    
+    # Test sorting by norm
+    sorted_indices_norm = cc._sort_planes(mode='norm')
+    assert sorted_indices_norm is not None
+    assert len(sorted_indices_norm) == 3
+    
+    # Test sorting by volume
+    sorted_indices_volume = cc._sort_planes(mode='volume')
+    assert sorted_indices_volume is not None
+    assert len(sorted_indices_volume) == 3
+    
+    # The sorted orders might differ between methods
+    # Largest volume should be first
+    volumes = np.prod(aabbs[:, 1, :] - aabbs[:, 0, :], axis=1)
+    assert sorted_indices_volume[0] == np.argmax(volumes)
+
+
+def test_inequalities():
+    """Test the inequalities conversion function."""
+    plane = np.array([1, 2, 3, -4])  # Plane equation: 1x + 2y + 3z - 4 = 0
+    
+    positive, negative = CellComplex._inequalities(plane)
+    
+    # Check format and values
+    assert len(positive) == 4
+    assert positive[0] == -4  # -4 + 1x + 2y + 3z >= 0
+    assert positive[1:4] == [1, 2, 3]
+    
+    assert len(negative) == 4
+    assert negative[0] == 4   # 4 - 1x - 2y - 3z >= 0
+    assert negative[1:4] == [-1, -2, -3]
+
+
+def test_cell_norm():
+    """Test cell normal calculation."""
+    # Create a simple rectangular set of vertices
+    vertices = np.array([
+        [0, 0, 0],
+        [1, 0, 0],
+        [1, 1, 0],
+        [0, 1, 0],
+        [0, 0, 1],
+        [1, 0, 1],
+        [1, 1, 1],
[…]

[…]
