### 2025-04-24 chenzhaiyu/abspy#32 @chenzhaiyu
Test coverage improved :)
```
Name                      Stmts   Miss  Cover
---------------------------------------------
abspy/__init__.py             6      0   100%
abspy/complex.py            495    152    69%
abspy/graph.py              208     76    63%
abspy/logger.py              22      8    64%
abspy/primitive.py          504    194    62%
abspy/version.py              3      1    67%
tests/test_complex.py       161      0   100%
tests/test_graph.py         162      0   100%
tests/test_primitive.py     312      0   100%
---------------------------------------------
TOTAL                      1873    431    77%
```
