### 2025-04-25 chenzhaiyu/abspy#32 @tuelwer
Hi @chenzhaiyu, 77% coverage looks good to me. When I run the tests, I get a couple of warnings:

```
tests/test_complex.py::test_visualise
tests/test_complex.py::test_visualise
  /opt/miniconda3/envs/abspy/lib/python3.11/ast.py:418: DeprecationWarning: visit_NameConstant is deprecated; add visit_Constant
    return visitor(node)

tests/test_primitive.py::test_quiet_parameter
  /Users/user/abspy-review/abspy/abspy/primitive.py:200: DeprecationWarning: string or file could not be read to its end due to unmatched data; this will raise a ValueError in the future.
    pc = np.fromstring(' '.join(pc_lines), sep=' ')

tests/test_primitive.py::test_fit_plane_with_lsa
  /Users/user/abspy-review/abspy/abspy/primitive.py:390: FutureWarning: `rcond` parameter will change to the default of machine precision times ``max(M, N)`` where M and N are the input matrix dimensions.
  To use the future default and silence this warning we advise to pass `rcond=None`, to keep using the old, explicitly pass `rcond=-1`.
    param = np.linalg.lstsq(points, np.expand_dims(np.ones(len(points)), 1))

-- Docs: https://docs.pytest.org/en/stable/how-to/capture-warnings.html
```
It might be a good idea to fix those.
