### 2025-04-25 chenzhaiyu/abspy#31 @tuelwer
Thanks for your response and your explanations. I still think there is some room for improvement regarding the explanations. The paper uses a lot of technical terminology which makes the paper difficult to read to folks not familiar with the topic. As these people are also potential users of abspy, the package would benefit from clearer explanations. Here are some ideas on how to do that:

- [ ] You could add an additional pane that shows the input into Figure 1.
- [ ] Add one or two sentences that explain what exactly is shown in Figure 2. Especially, the "points with associated planar primitives" and "representative sample points" are somewhat unclear to me.

Minor point: Please, improve the text flow. There is a lot of white space on the first page. Do lines 33-37 fit on the first page?
