### 2025-04-25 4298965
Replace deprecated `np.in1d` with `np.isin`
