### 2025-04-25 0606b86
Default `rcond=None` in `np.linalg.lstsq` (#32)
