### 2025-04-25 7a8d825
Fix VG input for `test_quiet_parameter` (#32)
