### 2025-04-25 3765942
Add input pane to Figure 1 (#31)
