### 2025-04-25 2520ac2
Improve Figure 2 caption and spacing (#31)
