+++ joss/paper.md
@@ -25,9 +25,10 @@ Efficient and robust space partitioning of 3D space underpins many applications 
 
 In many scientific and engineering workflows, raw 3D data must be transformed into compact volumetric models for visualization, simulation, and analysis of complex 3D environments. Traditional fixed-grid representations (e.g., voxels) or exhaustive plane-arrangement approaches [@edelsbrunner1986constructing] often either introduce discretization errors or incur prohibitive computational costs, or both. A central challenge is to partition ambient space into a minimal yet meaningful set of units that faithfully capture geometric detail. Adaptive binary space partitioning [@murali1997consistent] addresses this challenge by dynamically subdividing 3D space according to the data's intrinsic features, constructing a plane arrangement that conforms to geometric boundaries. Despite its theoretical advantages, no high-level implementation has been readily available for modern research use cases. *abspy* fills this gap by offering a modern, user-friendly API for adaptive partitioning, enabling reliable and efficient 3D modeling. Researchers can thus tackle large-scale and complex reconstruction challenges with great reliability and precision. Unlike general-purpose libraries such as CGAL [@fabri2009cgal] and Open3D [@zhou2018open3d], which primarily focus on versatile low-level tools, *abspy* uniquely integrates adaptive binary space partitioning and exact arithmetic within a single, Python-native framework, facilitating direct integration into research workflows tailored for compact, dynamic 3D modeling tasks and beyond.
 
+\autoref{fig:2d} illustrates how *abspy* uses user-supplied or extracted planar primitives to recursively subdivide space into a convex cell complex. Initially developed for research purposes, *abspy* has since evolved to support diverse 3D applications, particularly for compact surface reconstructions. It has been employed in several research projects and featured in scientific publications [@chen2022points2poly; @chen2024polygnn; @sulzer2024concise] as well as graduate students' projects.
+
 ![A 2D illustration for adaptive binary space partitioning. The ambient space is recursively partitioned into a cell complex with the insertion of planar primitives.\label{fig:2d}](assets/2d.png){ width=85% }
 
-\autoref{fig:2d} illustrates how *abspy* uses user-supplied or extracted planar primitives to recursively subdivide space into a convex cell complex. Initially developed for research purposes, *abspy* has since evolved to support diverse 3D applications, particularly for compact surface reconstructions. It has been employed in several research projects and featured in scientific publications [@chen2022points2poly; @chen2024polygnn; @sulzer2024concise] as well as graduate students' projects.
 
 # Overview of features
 
@@ -42,7 +43,7 @@ The core features of *abspy* include:
 
 \autoref{fig:3d} summarizes the primary representations supported by *abspy*. The [documentation of *abspy*](https://abspy.readthedocs.io/) contains tutorials for the package, its API reference, file format specifications, etc.
 
-![An overview of representations supported by *abspy*: points with associated planar primitives, adaptive cell complex, reconstructed surface mesh, representative sample points.\label{fig:3d}](assets/3d.png){ width=95% }
+![An overview of representations supported by *abspy*, shown left to right: points grouped and colored by planar primitives; cell complex of colored polyhedral volumes generated by adaptive space partitioning; reconstructed surface mesh; and sampled representative points from each cell volume, colored to match their cells.\label{fig:3d}](assets/3d.png){ width=95% }
 
 # Acknowledgements
 
