### 2025-04-25 28759ff
Try to improve Figure 2 spacing too (#31)
