+++ joss/paper.md
@@ -32,6 +32,8 @@ In many scientific and engineering workflows, raw 3D data must be transformed in
 
 # Overview of features
 
+![An overview of representations supported by *abspy*, shown left to right: points grouped and colored by planar primitives; cell complex of colored polyhedral volumes generated by adaptive space partitioning; reconstructed surface mesh; and sampled representative points from each cell volume, colored to match their cells.\label{fig:3d}](assets/3d.png){ width=95% }
+
 The core features of *abspy* include:
 
 - **Planar primitive manipulation:** *abspy* extracts planar primitives from point clouds or meshes, and refines or perturbs the primitives to control noise, ensuring a robust representation of the input geometric structures.
@@ -42,8 +44,6 @@ The core features of *abspy* include:
 - **Ease of integration:** With a Pythonic interface, comprehensive documentation, and practical examples, *abspy* integrates smoothly into existing research workflows. Its data structures are designed for interoperability with tools such as NetworkX [@hagberg2008exploring] and Easy3D [@nan2021easy3d], facilitating further development.
 
 \autoref{fig:3d} summarizes the primary representations supported by *abspy*. The [documentation of *abspy*](https://abspy.readthedocs.io/) contains tutorials for the package, its API reference, file format specifications, etc.
-
-![An overview of representations supported by *abspy*, shown left to right: points grouped and colored by planar primitives; cell complex of colored polyhedral volumes generated by adaptive space partitioning; reconstructed surface mesh; and sampled representative points from each cell volume, colored to match their cells.\label{fig:3d}](assets/3d.png){ width=95% }
 
 # Acknowledgements
 
