### 2025-04-26 chenzhaiyu/abspy#31 @tuelwer
Thanks for updating the caption for Figure 2. Overall, I still think that there is room improvement, especially in terms of background and problem definition. The problem that abspy solves is only discussed in Figure 1 and 2, but the explanations in the main text are still very limited. Can you please add more explanations (also in the main text)? Not all potential readers are familiar with the topic. Sorry for pushing back so hard but I think that this would really improve the paper.

Regarding Figure placement: In my opinion two figures per page is not optimal. Please, make sure that Figure 2 ends on page 3. This will likely be solved by implementing above requested changes.
