### 2025-04-27 47d9d94
Improve paper all around (#31)
