+++ joss/paper.md
@@ -19,20 +19,21 @@ bibliography: paper.bib
 
 # Summary
 
-Efficient and robust space partitioning of 3D space underpins many applications in computer graphics. *abspy* is a Python package for adaptive binary space partitioning and 3D modeling. At its core, *abspy* constructs a plane arrangement by recursively subdividing 3D space with planar primitives to form a linear cell complex that reflects the underlying geometric structure. This adaptive scheme iteratively refines the spatial decomposition, producing a compact representation that supports efficient 3D modeling and analysis. Built on robust computational techniques and interoperable data structures, *abspy* provides a flexible framework for advanced operations such as surface reconstruction, volumetric analysis, and feature extraction for machine learning.
+Efficient and robust space partitioning of 3D space underpins many applications in computer graphics. *abspy* is a Python package for adaptive binary space partitioning and 3D modeling. At its core, *abspy* constructs a plane arrangement by recursively subdividing 3D space with planar primitives to form a linear cell complex that reflects the underlying geometric structure. This adaptive scheme iteratively refines the spatial decomposition, producing a compact representation that supports efficient 3D modeling and analysis. Built on robust arithmetic kernels and interoperable data structures, *abspy* exposes a high‑level, Python‑native API that lets researchers prototype advanced 3D pipelines such as surface reconstruction, volumetric analysis, and feature extraction for machine learning.
 
 # Statement of need
 
-In many scientific and engineering workflows, raw 3D data must be transformed into compact volumetric models for visualization, simulation, and analysis of complex 3D environments. Traditional fixed-grid representations (e.g., voxels) or exhaustive plane-arrangement approaches [@edelsbrunner1986constructing] often either introduce discretization errors or incur prohibitive computational costs, or both. A central challenge is to partition ambient space into a minimal yet meaningful set of units that faithfully capture geometric detail. Adaptive binary space partitioning [@murali1997consistent] addresses this challenge by dynamically subdividing 3D space according to the data's intrinsic features, constructing a plane arrangement that conforms to geometric boundaries. Despite its theoretical advantages, no high-level implementation has been readily available for modern research use cases. *abspy* fills this gap by offering a modern, user-friendly API for adaptive partitioning, enabling reliable and efficient 3D modeling. Researchers can thus tackle large-scale and complex reconstruction challenges with great reliability and precision. Unlike general-purpose libraries such as CGAL [@fabri2009cgal] and Open3D [@zhou2018open3d], which primarily focus on versatile low-level tools, *abspy* uniquely integrates adaptive binary space partitioning and exact arithmetic within a single, Python-native framework, facilitating direct integration into research workflows tailored for compact, dynamic 3D modeling tasks and beyond.
+In many scientific and engineering workflows, raw 3D data must be transformed into compact volumetric representations for visualization, simulation, and analysis of complex 3D environments. Sensor-acquired point clouds and photogrammetric meshes often contain hundreds of millions of irregular, noisy samples that are expensive to store, transmit, and query, and that provide little inherent structure. Converting such data into a well-defined spatial decomposition, in which each cell represents a coherent geometric region, reduces memory usage, spares downstream algorithms from ad-hoc filtering and enables exact spatial predicates such as intersection, containment, and adjacency. The central challenge is to obtain this decomposition without inheriting the aliasing artefacts of coarse voxel grids or the combinatorial explosion produced by exhaustive plane arrangements [@edelsbrunner1986constructing].
 
[…]
