### 2025-04-27 e4e03fc
Remove extra punctuation (#31)
