+++ joss/paper.md
@@ -25,7 +25,7 @@ Efficient and robust space partitioning of 3D space underpins many applications 
 
 In many scientific and engineering workflows, raw 3D data must be transformed into compact volumetric representations for visualization, simulation, and analysis of complex 3D environments. Sensor-acquired point clouds and photogrammetric meshes often contain hundreds of millions of irregular, noisy samples that are expensive to store, transmit, and query, and that provide little inherent structure. Converting such data into a well-defined spatial decomposition, in which each cell represents a coherent geometric region, reduces memory usage, spares downstream algorithms from ad-hoc filtering and enables exact spatial predicates such as intersection, containment, and adjacency. The central challenge is to obtain this decomposition without inheriting the aliasing artefacts of coarse voxel grids or the combinatorial explosion produced by exhaustive plane arrangements [@edelsbrunner1986constructing].
 
-*How can 3D space be partitioned into a minimal yet meaningful set of units that faithfully capture geometric detail?*. Adaptive binary space partitioning (BSP) [@murali1997consistent] addresses this challenge by dynamically subdividing 3D space according to the data's intrinsic features, constructing a plane arrangement that conforms to geometric boundaries. Despite its theoretical advantages, a high-level implementation suitable for modern research use cases has been missing. *abspy* fills this gap with a modern, Python-native API for adaptive partitioning that delivers reliable and efficient 3D modeling. Researchers can thus tackle large-scale and complex reconstruction challenges with great reliability and precision. Unlike general-purpose libraries such as CGAL [@fabri2009cgal] and Open3D [@zhou2018open3d], which mainly supply low-level geometric utilities, *abspy* uniquely integrates adaptive BSP and exact arithmetic within a single, Python-native framework, facilitating direct integration into research workflows tailored for compact, dynamic 3D modeling tasks and beyond.
+*How can 3D space be partitioned into a minimal yet meaningful set of units that faithfully capture geometric detail?* Adaptive binary space partitioning (BSP) [@murali1997consistent] addresses this challenge by dynamically subdividing 3D space according to the data's intrinsic features, constructing a plane arrangement that conforms to geometric boundaries. Despite its theoretical advantages, a high-level implementation suitable for modern research use cases has been missing. *abspy* fills this gap with a modern, Python-native API for adaptive partitioning that delivers reliable and efficient 3D modeling. Researchers can thus tackle large-scale and complex reconstruction challenges with great reliability and precision. Unlike general-purpose libraries such as CGAL [@fabri2009cgal] and Open3D [@zhou2018open3d], which mainly supply low-level geometric utilities, *abspy* uniquely integrates adaptive BSP and exact arithmetic within a single, Python-native framework, facilitating direct integration into research workflows tailored for compact, dynamic 3D modeling tasks and beyond.
 
 \autoref{fig:2d} illustrates how *abspy* uses user-supplied or extracted planar primitives to recursively subdivide space into a convex cell complex. Beginning with user-supplied planes, we insert each plane only into the BSP-tree leaves whose bounding boxes it intersects. This local, on-demand splitting keeps every cell convex, updates adjacency on the fly, and, by skipping unrelated plane pairs, avoids the prohibitive cost of exhaustive arrangements. The result is a compact cell complex. Initially developed for research purposes, *abspy* has since evolved to support diverse 3D applications, particularly for compact surface reconstructions. It has been employed in several research projects and featured in scientific publications [@chen2022points2poly; @chen2024polygnn; @sulzer2024concise] as well as graduate students' projects.
 
