### 2025-04-28 chenzhaiyu/abspy#34 @tuelwer
Add instruction to run tests
Hi @chenzhaiyu, this is part of my https://github.com/openjournals/joss-reviews/issues/7946. Could you kindly add some brief instructions on how to run the tests? I know it is kind of obvious but I think it is good to have this explicitly stated somewhere as many developers would expect such instructions.
