### 2025-04-28 #7946 @tuelwer
Summary: abspy is a Python library for efficient 3D modeling using adaptive binary space partitioning. It enables accurate surface reconstruction and spatial analysis from point cloud data. The tool is designed for easy integration into research workflows.

General: The code base is licensed under MIT license and the submissing author has contributed a significant part of the codebase. Some [test data](https://github.com/chenzhaiyu/abspy/tree/main/tutorials/data) is shared and the submission meets the scope of eligibility for JOSS. The paper does not contain original results and no research on humans or animals has been done.

Functionality: The example code snippets demonstrate that the software works as intended. There are no performance claims in this paper. My only point of criticism is that the installation instructions heavily rely on conda which is in my opinion not optimal. Conda recently changed its licensing to a more permissive one which does not allow free use for commercial applications. I highly suggest to change the installation process (to use Python venvs or something similar) but I do not know whether this is easily implementable because of the sage dependency. This is not an acceptance blocker in my opinion.

Documentation: The repository contains [tutorials](https://abspy.readthedocs.io/en/latest/tutorials.html) which were improved in https://github.com/chenzhaiyu/abspy/issues/33. An extensive [API documentation](https://abspy.readthedocs.io/en/latest/api.html). There are automated [tests](https://github.com/chenzhaiyu/abspy/tree/main/tests) that cover 77% of the code base (see issue https://github.com/chenzhaiyu/abspy/issues/32 and https://github.com/chenzhaiyu/abspy/issues/34). Community guidelines can be found [here](https://github.com/chenzhaiyu/abspy/blob/main/CONTRIBUTING.md).

Software paper: The paper contains a summary and a statement of need. Related works are also discussed and the paper is well written. Some issues regarding the clarity of the paper were addressed by the author in https://github.com/chenzhaiyu/abspy/issues/31. abspy has been used in some recent journal and conference publications.

Very minor comment: The logo in the README.md is barely visible when Github is in dark-mode. I suggest adding a white background.

@matthewfeickert This is a solid contribution that is in my opinion suitable for publication in JOSS.
