### 2025-05-12 #7946 @matthewfeickert
@chenzhaiyu I'll open up a GitHub Issue with any real editorial comments, but as a quick one

>```
> 🟡 SKIP DOIs
>
> - No DOI given, and none found for title: Sage Mathematics Software (Version 10.2)
>```

Can you please use the information in the `CITATION.cff` for the SageMath citation (https://github.com/sagemath/sage/blob/94baf41dcd674ff894e0bb83a794d522281f5d7d/CITATION.cff)?
