### 2025-05-12 e2572a1
Update Sage citation with DOI
