+++ joss/paper.bib
@@ -78,13 +78,14 @@
   doi={10.1007/978-3-031-72904-1_21},
 }
 
-@manual{sage2024sage,
-  Key          = {Sage},
-  Author       = {W.\thinspace{}A. Stein and others},
-  Organization = {The Sage Development Team},
-  Title        = {{S}age {M}athematics {S}oftware ({V}ersion 10.2)},
-  note         = {{\tt http://www.sagemath.org}},
-  Year         = {2024},
+@software{sage2024sage,
+  author       = {The SageMath Developers},
+  title        = {{SageMath}},
+  month        = jan,
+  year         = 2025,
+  publisher    = {Zenodo},
+  version      = {10.6.beta3},
+  doi          = {10.5281/zenodo.8042260},
 }
 
 @inproceedings{fabri2009cgal,
