+++ joss/paper.bib
@@ -79,7 +79,7 @@
 }
 
 @software{sage2024sage,
-  author       = {The SageMath Developers},
+  author       = {{The SageMath Developers}},
   title        = {{SageMath}},
   month        = jan,
   year         = 2025,
