### 2025-05-12 #7946 @chenzhaiyu
> Can you please use the information in the `CITATION.cff` for the SageMath citation (https://github.com/sagemath/sage/blob/94baf41dcd674ff894e0bb83a794d522281f5d7d/CITATION.cff)?

Hi @matthewfeickert, thanks for the pointer! Fixed.
