### 2025-06-04 cbca8a6
macOS (both `osx-64` and `osx-arm64`) follows standard installation
