### 2025-06-05 chenzhaiyu/abspy#35 @matthewfeickert
[JOSS Review] Paper Editorial Suggestions
These are two very small editorial suggestions for the paper as part of https://github.com/openjournals/joss-reviews/issues/7946

* "exposes a high‑level, Python‑native API"

https://github.com/chenzhaiyu/abspy/blob/a5bbdeb12abb77af7dfd638c12fe171decc8a2a2/joss/paper.md?plain=1#L22

I'm not sure what "Python-native" means here. Unless there's an explicit need to clarify what is meant here in the text I'd recommend just using "Python API". There's other use of it later on

https://github.com/chenzhaiyu/abspy/blob/a5bbdeb12abb77af7dfd638c12fe171decc8a2a2/joss/paper.md?plain=1#L28

that I'd also suggest changing.

* "The documentation of abspy contains tutorials for the package, its API reference, file format specifications, etc."

https://github.com/chenzhaiyu/abspy/blob/a5bbdeb12abb77af7dfd638c12fe171decc8a2a2/joss/paper.md?plain=1#L47

"etc." isn't providing useful context here and should get dropped for something like

> The documentation of abspy contains tutorials for the package, its API reference, and file format specifications.

I think that the paper is in great shape and so I only have these nitpicky suggestions for you. 👍
