### 2025-06-05 b42fa7f
Add 'and' (#35)
