+++ joss/paper.md
@@ -44,7 +44,7 @@ The core features of *abspy* include:
 - **Surface reconstruction:** *abspy* identifies boundaries between interior and exterior cells to reconstruct polygonal surfaces using graph cuts, enabling the generation of surface models suitable for visualization and analysis.
 - **Ease of integration:** With a Pythonic interface, comprehensive documentation, and practical examples, *abspy* integrates smoothly into existing research workflows. Its data structures are designed for interoperability with tools such as NetworkX [@hagberg2008exploring] and Easy3D [@nan2021easy3d], facilitating further development.
 
-\autoref{fig:3d} showcases a few geometric representations that *abspy* can provide. From an input point cloud, for example, the library can manipulate points grouped by their supporting planar primitives, build a linear cell complex with its adaptive partitioning routine, extract a watertight polygonal surface from that complex through its graph-cut API, and sample representative points from the cells with the built-in samplers. These representations allow users to choose whichever level of abstraction best fits their application or to combine several of them as needed. The [documentation of *abspy*](https://abspy.readthedocs.io/) contains tutorials for the package, its API reference, file format specifications.
+\autoref{fig:3d} showcases a few geometric representations that *abspy* can provide. From an input point cloud, for example, the library can manipulate points grouped by their supporting planar primitives, build a linear cell complex with its adaptive partitioning routine, extract a watertight polygonal surface from that complex through its graph-cut API, and sample representative points from the cells with the built-in samplers. These representations allow users to choose whichever level of abstraction best fits their application or to combine several of them as needed. The [documentation of *abspy*](https://abspy.readthedocs.io/) contains tutorials for the package, its API reference, and file format specifications.
 
 ![An overview of representations supported by *abspy*, shown left to right: points grouped and colored by planar primitives; cell complex of colored polyhedral volumes generated by adaptive space partitioning; reconstructed surface mesh; and sampled representative points from each cell volume, colored to match their cells.\label{fig:3d}](assets/3d.png){ width=95% }
 
