+++ CHANGELOG.md
@@ -4,13 +4,19 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [0.3.0] - 2025-06-06
 ### Added
-- Unit tests integrated with GitHub Actions workflow.
+- Unit tests integrated with GitHub Actions workflow (Ubuntu and macOS).
+- Better test coverage (https://github.com/chenzhaiyu/abspy/issues/32).
+- More explanations for tutorials on the doc page (https://github.com/chenzhaiyu/abspy/issues/33).
 
 ### Changed
 - Rename previous `tests` to `tutorials`.
 - `CellComplex.prioritise_planes` defaults `prioritise_verticals=False`.
+- No global numpy print options (https://github.com/chenzhaiyu/abspy/issues/27).
+- Fix and complete `environment.yml`.
+- Fix obsolete benchmark (https://github.com/chenzhaiyu/abspy/issues/30).
+- Replace deprecated `np.in1d` with `np.isin`.
 
 ## [0.2.6] - 2025-02-03
 ### Added
