### 2025-06-06 0a2cf65
Bump version to 0.3.0
