+++ docs/source/conf.py
@@ -22,7 +22,7 @@ copyright = '2022, Zhaiyu Chen'
 author = 'Zhaiyu Chen'
 
 # The full version, including alpha/beta/rc tags
-release = '0.2.6'
+release = '0.3.0'
 
 
 # -- General configuration ---------------------------------------------------
