+++ setup.cfg
@@ -1,6 +1,6 @@
 [metadata]
 name = abspy
-version = 0.2.6
+version = 0.3.0
 author = Zhaiyu Chen
 author_email = zhaiyu.chen@outlook.com
 license='MIT'
