### 2025-06-06 #7946 @chenzhaiyu
-> https://github.com/chenzhaiyu/abspy/releases/tag/v0.3.0
-> [10.5281/zenodo.15609991](https://doi.org/10.5281/zenodo.15609991)

@matthewfeickert I've checked all the points. Let me know if I missed anything :)
