### 2025-06-06 #7946 @editorialbot
:wave: @openjournals/csism-eics, this paper is ready to be accepted and published.
