### 2025-06-06 #7946 @danielskatz
Congratulations to @chenzhaiyu (Zhaiyu Chen) on your JOSS publication!!

(It makes me happy when a submission gets to me for final acceptance and no changes are needed)
